Fixed some conversion-warnings.

+++ Frameworks/AIUtilities Framework/Source/NSCalendarDate+ISO8601Parsing.m
@@ -457,7 +457,7 @@ static BOOL is_leap_year(NSInteger year) {
 							if(*ch == '-') {
 								++ch;
 								month_or_week = read_segment_2digits(ch, &ch);
-								NSLog(@"(%@) month is %lu", str, month_or_week);
+								NSLog(@"(%@) month is %lu", str, (unsigned long)month_or_week);
 							}
 
 							day = 1U;
